# feat: Add K6 performance testing with web dashboard