--- a/.gitignore
+++ b/.gitignore
@@ -1,0 +1,3 @@
+node_modules
+results
+html-report.html