--- a/package.json
+++ b/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "sistema-estudos-server",
+  "version": "1.1.0",
+  "description": "Servidor local para Sistema de Controle de Estudos",
+  "scripts": {
+    "start": "node server.js",
+    "dev": "nodemon server.js",
+    "test:k6": "k6 run ./test/K6/performance.test.js",
+    "test-k6-with-export": "k6 run --summary-export ./test/K6/results/k6-results.json ./test/K6/performance.test.js",
+    "dash_k6": "set \"K6_WEB_DASHBOARD=true\" && set \"K6_WEB_DASHBOARD_EXPORT=html-report.html\" && k6 run ./test/K6/performance.test.js",
+    "test:cypress": "cypress run"
+  },
+  "dependencies": {
+    "express": "^4.18.2"
+  },
+  "devDependencies": {
+    "nodemon": "^3.0.1"
+  }
+}