- Add Node.js Express server for optimized performance
- Implement K6 performance tests with ramp-up scenarios
- Add web dashboard with real-time visualization
- Configure rigorous thresholds (P90<150ms, P95<300ms)
- Add NPM scripts for test automation
- Update documentation with K6 integration details
- Configure .gitignore for generated files

Performance results:
- 100% success rate with 100 concurrent users
- Zero failures in 3m35s stress test
- 20ms average response time, P95 < 75ms
- Throughput: 50+ req/s, 4 MB/s transfer rate